CLI reorg for DB configuration management operations
